Crate intrusive_splay_tree

Intrusive spaly tree

Structs

Node

A node of Tree. We made this public for the time being, expecting the need for the size of the left subtree when performing binary searches during insert or remove operations.

Tree

A splay tree.

Enums

Navi2

An enum used in a binary search that never stops, indicating whether to move left or right. This is used in insert.

Navi3

An enum indicating whether to proceed left or right during a binary search that might terminate early. This is used in remove.

Traits

Update

An adapter trait for specifying what to do during a structural change.